NCL-7411 Move PageTitles to common folder
