+++ src/components/ErrorPage/ErrorPage.tsx
@@ -1,10 +1,10 @@
 import { CubesIcon } from '@patternfly/react-icons';
 import { Link } from 'react-router-dom';
 
+import { PageTitles } from 'common/PageTitles';
+
 import { StateCard } from 'components/EmptyStates/StateCard';
 import { PageLayout } from 'components/PageLayout/PageLayout';
-
-import { PageTitles } from 'utils/PageTitles';
 
 import '../../index.css';
