+++ src/components/ProjectsPage/ProjectsPage.tsx
@@ -1,5 +1,7 @@
 import { Label } from '@patternfly/react-core';
 
+import { PageTitles } from 'common/PageTitles';
+
 import { useQueryParamsEffect } from 'hooks/useQueryParamsEffect';
 import { IService, useServiceContainer } from 'hooks/useServiceContainer';
 import { useTitle } from 'hooks/useTitle';
@@ -16,8 +18,6 @@ import { Toolbar } from 'components/Toolbar/Toolbar';
 import { ToolbarItem } from 'components/Toolbar/ToolbarItem';
 
 import * as projectApi from 'services/projectApi';
-
-import { PageTitles } from 'utils/PageTitles';
 
 // keeping also not supported operations for testing purposes
 const filterOptions: IFilterOptions = {
